OAuth/auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
jinja2
redis